Support PNPM
PNPM[pnpm] is NPM for developers with lots of projects and dependencies.

PNPM is a package manager alternative for NPM which uses multiple strategies to
prefer linking dependencies in your project from a single per-machine location.

PNPM, at this time of commit authoring, supports a server that can be started
and stopped. The server will, on its own, let you know when there are new
dependency versions available and a number of other things.

[pnpm]: https://pnpm.js.org "pnpm - Fast, disk space efficient npm installs"

Signed-off-by: Johnneylee Jack Rollins <Johnneylee.Rollins@gmail.com>